ORIGINALLY AUTHORED BY kurtamohler

Add `torch.storage.TypedStorage`, a wrapper for the untyped `ByteStorage` class, and use it as a base class to re-implement all the other `<type>Storage` classes.  `TypedStorage` has the same interface as `ByteStorage`. Serialization and other internals are updated accordingly, maintaining as much FC/BC as possible.
    
Fixes https://github.com/pytorch/pytorch/issues/47442
    
* **THE SERIALIZATION FORMAT IS FULLY FC/BC.** We worked very hard to make sure this is the case. We will probably want to break FC at some point to make the serialization structure of tensors make more sense, but not today.
* Storage classes have a new size method, `nbytes`.
* The existing interface for all the `<type>Storage` classes is maintained, except in a few cases, which now throw errors.
* `TypedStorage` objects can be instantiated directly and used interchangeably with their `<type>Storage` counterparts.
* `Storage._new_with_file`, `_set_from_file`, and `_write_file` now require explicit element size arguments.
* `ByteStorage.from_buffer` now requires a `dtype` kwarg.
* The preexisting serialization format stores dtype with storage, and in fact this dtype is used to determine the dtype of the tensor overall.
**If you overrode the handling of pickling/unpickling, you MUST add handling for TypedStorage** or your serialization code will degrade to standard file-based serialization.


Original pull request: https://github.com/pytorch/pytorch/pull/59671

Differential Revision: [D29466819](https://our.internmc.facebook.com/intern/diff/D29466819/)

cc pietern mrshenli pritamdamania87 zhaojuanmao satgera rohan-varma gqchen aazzolini osalpekar jiayisuse agolynski @SciPioneer @H-Huang mrzzd cbalioglu gcramer23

[ghstack-poisoned]

+++ torch/distributed/distributed_c10d.py
@@ -1518,7 +1518,7 @@ def all_gather_multigpu(
 def _object_to_tensor(obj):
     f = io.BytesIO()
     _pickler(f).dump(obj)
-    byte_storage = torch.ByteStorage.from_buffer(f.getvalue())  # type: ignore[attr-defined]
+    byte_storage = torch.ByteStorage.from_buffer(f.getvalue(), dtype=torch.uint8)  # type: ignore[attr-defined]
     byte_tensor = torch.tensor(byte_storage, dtype=torch.uint8)
     local_size = torch.tensor([byte_tensor.numel()], dtype=torch.long)
     return byte_tensor, local_size
